# backend/tests/contract/snapshots/api-v1.openapi.json
{
  "components": {
    "schemas": {
...
      "Body_QI-API-PRJ-001": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_QI-API-PRJ-001",
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "ProjectListResponse": {
        "additionalProperties": false,
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProjectListItemResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items",
          "count"
        ],
        "title": "ProjectListResponse",
        "type": "object"
      },
      "ProjectPhase": {
        "enum": [
          "queued",
          "processing",
          "ready_for_review",
          "partial_review_required",
          "failed"
        ],
        "title": "ProjectPhase",
        "type": "string"
      },
...
      "ProjectStatusResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectError"
              },
              {
                "type": "null"
              }
            ]
          },
          "phase": {
            "$ref": "#/components/schemas/ProjectPhase"
          },
          "project_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          },
          "stage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProcessingStage"
              },
              {
                "type": "null"
              }
            ]
          },
          "workbench_ready": {
            "title": "Workbench Ready",
            "type": "boolean"
          }
        },
        "required": [
          "phase",
          "workbench_ready",
          "retryable",
          "error"
        ],
        "title": "ProjectStatusResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Quality Inspection",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/projects": {
      "get": {
        "operationId": "QI-API-PRJ-006",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unexpected internal failure.",
            "x-stable-error-codes": [
              "project_list_failed",
              "internal_server_error"
            ]
          }
        },
        "summary": "List Projects",
        "tags": [
          "projects"
        ]
      },
      "post": {
        "operationId": "QI-API-PRJ-001",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_QI-API-PRJ-001"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request failed transport or business validation.",
            "x-stable-error-codes": [
              "invalid_pdf",
              "request_validation_failed"
            ]
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unexpected internal failure.",
            "x-stable-error-codes": [
              "project_intake_failed",
              "internal_server_error"
            ]
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Required dispatch or service is unavailable.",
            "x-stable-error-codes": [
              "project_dispatch_failed"
            ]
          }
        },
        "summary": "Create Project",
        "tags": [
          "projects"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/review/commands": {
      "post": {
        "operationId": "QI-API-REV-003",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-QI-Operator",
            "required": true,
            "schema": {
              "title": "X-Qi-Operator",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewCommandRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewWorkingCopyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Requested resource was not found.",
            "x-stable-error-codes": [
              "project_not_found",
              "review_working_copy_not_found"
            ]
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request conflicts with current aggregate or capability state.",
            "x-stable-error-codes": [
              "review_version_conflict",
              "review_items_frozen",
              "review_lock_conflict"
            ]
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request failed transport or business validation.",
            "x-stable-error-codes": [
              "review_command_invalid",
              "request_validation_failed"
            ]
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unexpected internal failure.",
            "x-stable-error-codes": [
              "internal_server_error"
            ]
          }
        },
        "summary": "Apply Command",
        "tags": [
          "review"
        ]
      }
    },
    "/api/v1/projects/{project_id}/review/confirm": {
      "post": {
        "operationId": "QI-API-REV-005",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-QI-Operator",
            "required": true,
            "schema": {
              "title": "X-Qi-Operator",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewedResultResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Requested resource was not found.",
            "x-stable-error-codes": [
              "project_not_found",
              "review_working_copy_not_found"
            ]
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request conflicts with current aggregate or capability state.",
            "x-stable-error-codes": [
              "review_version_conflict",
              "item_set_not_frozen",
              "numbering_stale",
              "sip_metadata_unconfirmed",
              "sip_detail_fields_unconfirmed",
              "missing_required_balloon",
              "item_balloon_disconnect",
              "manual_required",
              "unreadable_number",
              "outside_cropbox",
              "owner_glyph_outside_circle",
              "protected_overlap",
              "source_text_overlap",
              "invalid_leader",
              "circle_overlap",
              "glyph_overlap",
              "glyph_circle_overlap",
              "duplicate_or_gapped_number",
              "review_lock_conflict"
            ]
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request failed transport or business validation.",
            "x-stable-error-codes": [
              "review_confirmation_invalid",
              "request_validation_failed"
            ]
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unexpected internal failure.",
            "x-stable-error-codes": [
              "internal_server_error"
            ]
          }
        },
        "summary": "Confirm Review",
        "tags": [
          "review"
        ]
      }
    },
    "/api/v1/projects/{project_id}/review/freeze-items": {
      "post": {
        "operationId": "QI-API-REV-004",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-QI-Operator",
            "required": true,
            "schema": {
              "title": "X-Qi-Operator",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FreezeItemsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewWorkingCopyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Requested resource was not found.",
            "x-stable-error-codes": [
              "project_not_found",
              "review_working_copy_not_found"
            ]
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request conflicts with current aggregate or capability state.",
            "x-stable-error-codes": [
              "review_version_conflict",
              "coverage_blocking",
              "unresolved_confirmation",
              "balloon_required_unconfirmed",
              "review_items_frozen",
              "review_lock_conflict"
            ]
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request failed transport or business validation.",
            "x-stable-error-codes": [
              "review_operator_invalid",
              "request_validation_failed"
            ]
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unexpected internal failure.",
            "x-stable-error-codes": [
              "internal_server_error"
            ]
          }
        },
        "summary": "Freeze Items",
        "tags": [
          "review"
        ]
      }
    },
    "/api/v1/projects/{project_id}/review/lock": {
      "post": {
        "operationId": "QI-API-REV-001",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-QI-Operator",
            "required": true,
            "schema": {
              "title": "X-Qi-Operator",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LockRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewLockResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Requested resource was not found.",
            "x-stable-error-codes": [
              "project_not_found"
            ]
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request conflicts with current aggregate or capability state.",
            "x-stable-error-codes": [
              "review_lock_conflict"
            ]
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request failed transport or business validation.",
            "x-stable-error-codes": [
              "request_validation_failed"
            ]
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unexpected internal failure.",
            "x-stable-error-codes": [
              "internal_server_error"
            ]
          }
        },
        "summary": "Lock Review",
        "tags": [
          "review"
        ]
      }
    },
    "/api/v1/projects/{project_id}/review/lock/release": {
      "post": {
        "operationId": "QI-API-REV-006",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-QI-Operator",
            "required": true,
            "schema": {
              "title": "X-Qi-Operator",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReleaseLockRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewLockReleaseResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Requested resource was not found.",
            "x-stable-error-codes": [
              "project_not_found"
            ]
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request failed transport or business validation.",
            "x-stable-error-codes": [
              "review_operator_invalid",
              "request_validation_failed"
            ]
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unexpected internal failure.",
            "x-stable-error-codes": [
              "internal_server_error"
            ]
          }
        },
        "summary": "Release Review Lock",
        "tags": [
          "review"
        ]
      }
    },
    "/api/v1/projects/{project_id}/review/working-copy": {
      "get": {
        "operationId": "QI-API-REV-002",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewWorkingCopyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Requested resource was not found.",
            "x-stable-error-codes": [
              "project_not_found",
              "review_working_copy_not_found"
            ]
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request failed transport or business validation.",
            "x-stable-error-codes": [
              "request_validation_failed"
            ]
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unexpected internal failure.",
            "x-stable-error-codes": [
              "internal_server_error"
            ]
          }
        },
        "summary": "Get Working Copy",
        "tags": [
          "review"
        ]
      }
    },
...
  }
}